Remove unnecessary sos_write just for stdout and link tty to libsosapi the same as sosh
