--- apps/tty_test/src/ttyout.c
@@ -1,60 +1,0 @@
-/*
- * Copyright 2019, Data61
- * Commonwealth Scientific and Industrial Research Organisation (CSIRO)
- * ABN 41 687 119 230.
- *
- * This software may be distributed and modified according to the terms of
- * the GNU General Public License version 2. Note that NO WARRANTY is provided.
- * See "LICENSE_GPLv2.txt" for details.
- *
- * @TAG(DATA61_GPL)
- */
-/****************************************************************************
- *
- *      $Id:  $
- *
- *      Description: Simple milestone 0 code.
- *                   Libc will need sos_write & sos_read implemented.
- *
- *      Author:      Ben Leslie
- *
- ****************************************************************************/
-
-#include <stdarg.h>
-#include <assert.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "ttyout.h"
-
-#include <sel4/sel4.h>
-
-void ttyout_init(void)
-{
-    /* Perform any initialisation you require here */
-}
-
-static size_t sos_debug_print(const void *vData, size_t count)
-{
-#ifdef CONFIG_DEBUG_BUILD
-    size_t i;
-    const char *realdata = vData;
-    for (i = 0; i < count; i++) {
-        seL4_DebugPutChar(realdata[i]);
-    }
-#endif
-    return count;
-}
-
-size_t sos_write(void *vData, size_t count)
-{
-    //implement this to use your syscall
-    return sos_debug_print(vData, count);
-}
-
-size_t sos_read(void *vData, size_t count)
-{
-    //implement this to use your syscall
-    return 0;
-}
-
